api- updates and email responder

+++ aws-ace-covid19/todos/delete.py
@@ -1,4 +1,5 @@
 import os
+import hashlib
 
 import boto3
 dynamodb = boto3.resource('dynamodb')
@@ -12,6 +13,25 @@ def delete(event, context):
         raise Exception("Couldn't create the todo item.")
 
     table = dynamodb.Table(os.environ['DYNAMODB_TABLE'])
+
+    md = hashlib.md5();
+    vCrc = data['crc']
+    vTst = str(data['uid']+'yabba').encode('utf-8');
+    md.update(vTst)
+    vHash = md.hexdigest()
+    
+    if(vCrc != vHash):
+        response = {
+            "statusCode": 502,
+            "headers": {
+                "Access-Control-Allow-Origin":"*",
+                "Access-Control-Allow-Methods":"*",
+                "Access-Control-Allow-Headers":"*"
+            },
+            "body": ""
+        }
+        return response
+
 
     # delete the todo from the database
     table.delete_item(
